+++ algorithms/src/shortest_paths/min_scored.rs
@@ -1,0 +1,61 @@
+use core::cmp::Ordering;
+
+pub trait TotalOrd {
+    fn total_cmp(&self, other: &Self) -> Ordering;
+}
+
+impl<T> TotalOrd for T
+where
+    T: Ord,
+{
+    fn total_cmp(&self, other: &Self) -> Ordering {
+        self.cmp(other)
+    }
+}
+
+impl TotalOrd for f32 {
+    fn total_cmp(&self, other: &Self) -> Ordering {
+        self.total_cmp(other)
+    }
+}
+
+impl TotalOrd for f64 {
+    fn total_cmp(&self, other: &Self) -> Ordering {
+        self.total_cmp(other)
+    }
+}
+
+/// `MinScored<K, T>` holds a score `K` and a scored object `T` in
+/// a pair for use with a `BinaryHeap`.
+///
+/// `MinScored` compares in reverse order by the score, so that we can
+/// use `BinaryHeap` as a min-heap to extract the score-value pair with the
+/// least score.
+///
+/// **Note:** `MinScored` implements a total order (`Ord`), so that it is
+/// possible to use float types as scores.
+#[derive(Copy, Clone, Debug)]
+pub struct MinScored<K, T>(pub K, pub T);
+
+impl<K: TotalOrd, T> PartialEq for MinScored<K, T> {
+    #[inline]
+    fn eq(&self, other: &MinScored<K, T>) -> bool {
+        self.cmp(other) == Ordering::Equal
+    }
+}
+
+impl<K: TotalOrd, T> Eq for MinScored<K, T> {}
+
+impl<K: TotalOrd, T> PartialOrd for MinScored<K, T> {
+    #[inline]
+    fn partial_cmp(&self, other: &MinScored<K, T>) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl<K: TotalOrd, T> Ord for MinScored<K, T> {
+    #[inline]
+    fn cmp(&self, other: &MinScored<K, T>) -> Ordering {
+        self.0.total_cmp(&other.0)
+    }
+}