+++ algorithms/src/shortest_paths/mod.rs
@@ -1,3 +1,13 @@
-pub(crate) mod astar;
+mod astar;
+mod bellman_ford;
+mod dijkstra;
+mod k_shortest_paths;
+mod measure;
+mod min_scored;
 
 pub use astar::astar;
+pub use bellman_ford::{bellman_ford, find_negative_cycle};
+pub use dijkstra::dijkstra;
+pub use k_shortest_paths::k_shortest_paths;
+pub use measure::{BoundedMeasure, FloatMeasure, Measure};
+pub use min_scored::TotalOrd;